Fix SpeedHistory overflow (uint8→uint16); reduce tint multiplier 10→4
uint8_t clamped all speeds >255 to 255, making 256x/512x/8192x look
identical — no change lines were drawn above 255x. uint16_t covers 8192.
Tint multiplier reduced so background stays subtle across the full range.

+++ econ/market.h
@@ -10,7 +10,7 @@
 // Circular buffer tracking sim speed (ticks_per_frame) at each recorded sample,
 // so price-history plots can overlay tick marks showing relative simulation speed.
 typedef struct {
-    uint8_t data[PRICE_HISTORY_SIZE];
+    uint16_t data[PRICE_HISTORY_SIZE];
     int     count;
     int     head;
     int     total_frames;     // total frames ever recorded (monotonically increasing; reset on clear)
